@@ -11,41 +11,41 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "~10.0.2",
-    "@angular/cdk": "^10.0.1",
-    "@angular/common": "~10.0.2",
-    "@angular/compiler": "~10.0.2",
-    "@angular/core": "~10.0.2",
-    "@angular/flex-layout": "^10.0.0-beta.32",
-    "@angular/forms": "~10.0.2",
-    "@angular/material": "^10.0.1",
-    "@angular/platform-browser": "~10.0.2",
-    "@angular/platform-browser-dynamic": "~10.0.2",
-    "@angular/router": "~10.0.2",
-    "core-js": "^3.6.5",
-    "rxjs": "~6.6.0",
-    "tslib": "^2.0.0",
-    "zone.js": "~0.10.3"
+    "@angular/animations": "~11.0.2",
+    "@angular/cdk": "^11.0.1",
+    "@angular/common": "~11.0.2",
+    "@angular/compiler": "~11.0.2",
+    "@angular/core": "~11.0.2",
+    "@angular/flex-layout": "^11.0.0-beta.33",
+    "@angular/forms": "~11.0.2",
+    "@angular/material": "^11.0.1",
+    "@angular/platform-browser": "~11.0.2",
+    "@angular/platform-browser-dynamic": "~11.0.2",
+    "@angular/router": "~11.0.2",
+    "core-js": "^3.8.0",
+    "rxjs": "~6.6.3",
+    "tslib": "^2.0.3",
+    "zone.js": "~0.11.3"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "^0.1000.1",
-    "@angular/cli": "~10.0.1",
-    "@angular/compiler-cli": "~10.0.2",
-    "@angular/language-service": "~10.0.2",
-    "@types/jasmine": "~3.5.11",
+    "@angular-devkit/build-angular": "^0.1100.2",
+    "@angular/cli": "~11.0.2",
+    "@angular/compiler-cli": "~11.0.2",
+    "@angular/language-service": "~11.0.2",
+    "@types/jasmine": "~3.6.2",
     "@types/jasminewd2": "~2.0.8",
-    "@types/node": "^14.0.14",
-    "codelyzer": "^6.0.0",
-    "jasmine-core": "~3.5.0",
-    "jasmine-spec-reporter": "~5.0.2",
-    "karma": "~5.1.0",
+    "@types/node": "^14.14.10",
+    "codelyzer": "^6.0.1",
+    "jasmine-core": "~3.6.0",
+    "jasmine-spec-reporter": "~6.0.0",
+    "karma": "~5.2.3",
     "karma-chrome-launcher": "~3.1.0",
     "karma-coverage-istanbul-reporter": "~3.0.3",
-    "karma-jasmine": "~3.3.1",
+    "karma-jasmine": "~4.0.1",
     "karma-jasmine-html-reporter": "^1.5.4",
     "protractor": "~7.0.0",
-    "ts-node": "~8.10.2",
+    "ts-node": "~9.0.0",
     "tslint": "~6.1.2",
-    "typescript": "3.9.6"
+    "typescript": "4.0.2"
   }
 }
